Moved the creative tab pointers list to Utils Makes more sense to have it there since we'll reference it from blocks too.
