+++ src/main/java/net/madelyn/nyagibits_bytes/misc/Utils.java
@@ -1,7 +1,9 @@
 package net.madelyn.nyagibits_bytes.misc;
 
 import com.mojang.logging.LogUtils;
+import net.madelyn.nyagibits_bytes.item.ModCreativeModeTab;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.Block;
@@ -30,5 +32,17 @@ public class Utils {
         }
         return block;
     }
+
+    //This is just a list of shorthands to make the list cleaner.
+    public static class Tab{
+        public static CreativeModeTab ITEMS = ModCreativeModeTab.NYAGIBITS_BYTES_ITEMS;
+        public static CreativeModeTab BLOCKS = ModCreativeModeTab.NYAGIBITS_BYTES_BLOCKS;
+        public static CreativeModeTab SCIENCE = ModCreativeModeTab.NYAGIBITS_BYTES_SCIENCE;
+        public static CreativeModeTab SCHEMATICS = ModCreativeModeTab.NYAGIBITS_BYTES_SCHEMATICS;
+        public static CreativeModeTab SEQUENCE = ModCreativeModeTab.NYAGIBITS_BYTES_INCOMPLETE_SEQUENCE_ITEMS;
+        public static CreativeModeTab FLUIDS = ModCreativeModeTab.NYAGIBITS_BYTES_FLUIDS;
+        public static CreativeModeTab MINERALS = ModCreativeModeTab.NYAGIBITS_BYTES_MINERALS;
+    }
+
 
 }
